+++ packages/components/typography/src/index.ts
@@ -1,4 +1,4 @@
-import * as macro from './style/macro';
+import * as macro from './style/macro-utils';
 
 export { default } from './Typography';
 export * from './CommonElements';