+++ packages/portal/stories/components/Radio/index.stories.tsx
@@ -25,7 +25,7 @@ const stories = {
     </Radio.Group>
   ),
   radioButtons: () => (
-    <Radio.Group defaultValue="a" buttonStyle="solid" onChange={action('onChange')}>
+    <Radio.Group defaultValue="a" buttonStyle="solid" onChange={action('onChange')} fullWidth={boolean('fullWidth', false)}>
       <Radio.Button value="a">A</Radio.Button>
       <Radio.Button value="b" disabled={boolean('disable B', false)}>B</Radio.Button>
       <Radio.Button value="c">C</Radio.Button>