+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.4.0...@synerise/ds-icon@0.5.0) (2019-11-22)
+
+### Features
+
+- add new icons ([7a02147](https://github.com/Synerise/synerise-design/commit/7a021474f49ffe4e7d7582735e33bd52e7b4f518))
+
+# [0.4.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.3.5...@synerise/ds-icon@0.4.0) (2019-11-18)
+
+### Features
+
+- **icon:** Fix icons names ([85f0cdd](https://github.com/Synerise/synerise-design/commit/85f0cddb0306a63ed2dfefb6b8edd6ad734a3ee5))
+
+## [0.3.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.3.3...@synerise/ds-icon@0.3.4) (2019-11-15)
+
+### Bug Fixes
+
+- **icon:** remove style colors from svg ([f196404](https://github.com/Synerise/synerise-design/commit/f19640446c89f4e15c2fa493e65d05a43990f988))
+
 ## [0.3.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.3.1...@synerise/ds-icon@0.3.2) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-icon