+++ packages/components/radio/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-radio@0.6.0...@synerise/ds-radio@0.7.0) (2019-11-22)
+
+### Features
+
+- **radio:** Fix radio button style ([f3d3d27](https://github.com/synerise/synerise-design/commit/f3d3d276901b058ecde7687ff3eff31b2e4a654e))
+
+# [0.6.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-radio@0.5.6...@synerise/ds-radio@0.6.0) (2019-11-22)
+
+### Features
+
+- **radio:** Enhance radio button ([343ab5b](https://github.com/synerise/synerise-design/commit/343ab5b3f7fc4e7f6d95084df941155823f7a14e))
+
 ## [0.5.6](https://github.com/synerise/synerise-design/compare/@synerise/ds-radio@0.5.5...@synerise/ds-radio@0.5.6) (2019-11-09)
 
 **Note:** Version bump only for package @synerise/ds-radio