+++ packages/components/status/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.19](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.18...@synerise/ds-status@0.0.19) (2019-11-22)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.18](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.17...@synerise/ds-status@0.0.18) (2019-11-18)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.16](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.15...@synerise/ds-status@0.0.16) (2019-11-18)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.15](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.14...@synerise/ds-status@0.0.15) (2019-11-15)
+
+**Note:** Version bump only for package @synerise/ds-status
+
 ## [0.0.13](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.12...@synerise/ds-status@0.0.13) (2019-11-09)
 
 **Note:** Version bump only for package @synerise/ds-status